Sửa lại phần xóa

+++ src/app/components/delete-tour/delete-tour.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core';
+import {Component, OnInit} from '@angular/core';
 import {FormControl, FormGroup} from "@angular/forms";
 import {Tour} from "../../models/tour";
 import {TourService} from "../../services/tour.service";
@@ -10,29 +10,20 @@ import {ActivatedRoute, Router} from "@angular/router";
   styleUrls: ['./delete-tour.component.css']
 })
 export class DeleteTourComponent implements OnInit {
-  deleteForm : FormGroup = new FormGroup({
-      title: new FormControl(),
-      price: new FormControl(),
-      description: new FormControl(),
-    }
-  )
 
-  id : any;
-  tour : Tour | any;
+  id: any;
+  tour: Tour | any;
 
-  constructor(private tourService : TourService,
-              private router : Router,
-              private activatedRoute : ActivatedRoute) { }
+  constructor(private tourService: TourService,
+              private router: Router,
+              private activatedRoute: ActivatedRoute) {
+  }
 
   ngOnInit(): void {
     this.activatedRoute.paramMap.subscribe(parammap => {
       this.id = parammap.get('id');
       this.tourService.getById(this.id).subscribe(data => {
-          this.deleteForm.patchValue({
-            title: data.title,
-            price: data.price,
-            description: data.description
-          })
+          this.tour = data;
         },
         error => {
           console.log(error);
@@ -41,12 +32,16 @@ export class DeleteTourComponent implements OnInit {
   }
 
   deleteTour() {
-    this.tourService.deleteTour(this.id).subscribe(() => {
-      this.router.navigateByUrl("/")
-      alert("Đã xóa xong!!")
-    }, error => {
-      console.log(error)
-    })
+    if (confirm("Bạn có muốn xóa không?")) {
+      this.tourService.deleteTour(this.id).subscribe(() => {
+        this.router.navigateByUrl("/");
+        alert("Đã xóa xong!!")
+      }, error => {
+        console.log(error)
+      });
+    } else {
+      this.router.navigateByUrl("/delete/" + this.id);
+    }
   }
 
 }
